refactor: record create view로 변경
